refactor : SJoinForm -> SForm
