+++ src/styles/components/index.ts
@@ -1,2 +1,2 @@
+export * from './common';
 export * from './flex';
-export * from './join';
